requirements.txt
numpy==1.24.2
pandas==2.0.0
PyYAML==6.0.1
scikit-learn==1.2.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0